refactor(drawer): move files out of internal folder
