+++ src/components/dialog-full-screen/dialog-full-screen.stories.mdx
@@ -9,7 +9,7 @@ import Button from "../button";
 import Form from "../form";
 import Textbox from "../../__experimental__/components/textbox";
 import Pill from "../../components/pill";
-import Drawer from "../drawer/__internal__/drawer.component";
+import Drawer from "../drawer/drawer.component";
 import { Tabs, Tab } from "../tabs";
 import useMediaQuery from "../../hooks/useMediaQuery";
 import Link from "../link";
