+++ src/components/drawer/drawer.component.js
@@ -2,8 +2,8 @@ import React, { useState, useCallback, useRef, useEffect } from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
 import invariant from "invariant";
-import createGuid from "../../../utils/helpers/guid/guid";
-import Icon from "../../icon";
+import createGuid from "../../utils/helpers/guid/guid";
+import Icon from "../icon";
 
 import {
   StyledDrawerWrapper,
