+++ src/components/drawer/drawer.style.js
@@ -1,6 +1,6 @@
 import styled, { css, keyframes } from "styled-components";
-import baseTheme from "../../../style/themes/base";
-import StyledTabs from "../../tabs/tabs.style";
+import baseTheme from "../../style/themes/base";
+import StyledTabs from "../tabs/tabs.style";
 
 const StyledSidebarTitle = styled.div`
   margin: 24px 36px 24px 24px;
