+++ src/components/drawer/index.d.ts
@@ -1,1 +1,1 @@
-export { default, SidebarContext } from "./__internal__/drawer";
+export { default, SidebarContext } from "./drawer";
